+++ library-domain/src/main/kotlin/dev/brevitz/nike/library/domain/player/PlayerRepository.kt
@@ -3,5 +3,5 @@ package dev.brevitz.nike.library.domain.player
 import dev.brevitz.nike.core.domain.ObservableRemoteData
 
 interface PlayerRepository {
-    fun getDetails(id: Int): ObservableRemoteData<Player>
+    fun getDetails(id: Long): ObservableRemoteData<Player>
 }
