+++ library-domain/src/main/kotlin/dev/brevitz/nike/library/domain/player/PlayerStore.kt
@@ -2,4 +2,4 @@ package dev.brevitz.nike.library.domain.player
 
 import dev.brevitz.nike.core.domain.ReactiveStore
 
-typealias PlayerStore = ReactiveStore<Int, Player>
+typealias PlayerStore = ReactiveStore<Long, Player>
